feat: adding Role CRUD and assigning permission to role. Creating DTO as well

+++ Services/RoleService.cs
@@ -8,10 +8,10 @@ namespace CMC.TS.FT.Api.Services
 {
     public class RoleService
     {
-        private readonly ILogger _logger;
+        private readonly ILogger<RoleService> _logger;
         private readonly IRoleRepository _roleRepository;
 
-        public RoleService(ILogger logger, IRoleRepository roleRepository)
+        public RoleService(ILogger<RoleService> logger, IRoleRepository roleRepository)
         {
             _roleRepository = roleRepository;
             _logger = logger;
@@ -22,22 +22,46 @@ namespace CMC.TS.FT.Api.Services
         {
             try
             {
-                _logger.LogInformation("Assign permission to role operation start");
+                //Log bắt đầu
+                _logger.LogInformation("SyncRolePermission operation start");
+
+                //PermissionIds không được phép null
                 if(rolePermission.PermissionIds == null)
                 {
                     _logger.LogError("permission Id can't be null");
                     return false;
                 }
-
+                //nếu PermissionIds hoàn toàn trống. tức người dùng chỉ muốn xoá. Thì xoá và thoát.
                 //xoá hết permission cũ của role.
-                bool isSuccess = await _roleRepository.DeleteRolePermissionByRoleId(rolePermission.RoleId);
-                if(isSuccess == false)
+                if(rolePermission.PermissionIds.Count == 0)
                 {
-                    _logger.LogInformation("Can't delete permission");
+                    bool isSuccess = await _roleRepository.DeleteRolePermissionByRoleId(rolePermission.RoleId);
+                    if (isSuccess == false)
+                    {
+                        _logger.LogInformation("Can't delete permission");
+                        return false;
+                    }
+                    else if (isSuccess == true)
+                    {
+                        _logger.LogInformation("delete permission success");
+                        return true;
+                    }
                 }
 
+                //nếu permissionIds không hoàn toàn trống. Thì vẫn xoá hết permissionId trong database và thêm cái mới vào.
                 if (rolePermission.PermissionIds.Count != 0)
                 {
+                    //xoá các permissionId cũ
+                    bool isSuccess = await _roleRepository.DeleteRolePermissionByRoleId(rolePermission.RoleId);
+                    if (isSuccess == false)
+                    {
+                        _logger.LogInformation("Can't delete permission");
+                    }
+                    else if (isSuccess == true)
+                    {
+                        _logger.LogInformation("delete permission success");
+                    }
+
                     //Thêm các permission mới vào role
                     List<RolePermission> thisisnew = rolePermission.PermissionIds.Select(pi => new RolePermission
                     {
@@ -78,6 +102,108 @@ namespace CMC.TS.FT.Api.Services
             catch (Exception e)
             {
                 _logger.LogError(e, "Assign permission to role operation failed");
+                return false;
+            }
+        }
+        public async Task<bool> CreateNewRole(CreateRoleDTO createRole)
+        {
+            try
+            {
+                _logger.LogInformation("create new role operation start");
+                Role role = new Role 
+                { 
+                    RoleId = Guid.NewGuid(),
+                    RoleName = createRole.RoleName,
+                    CreateAt = DateTime.UtcNow,
+                    UpdateAt = null,
+                    CreateBy = Guid.Empty,
+                    UpdateBy = Guid.Empty
+                };
+
+                bool isSuccess = await _roleRepository.Create(role);
+                if(isSuccess == true)
+                {
+                    _logger.LogInformation("create new role success");
+                    return true;
+                }
+                else
+                {
+                    _logger.LogInformation("create new role failed");
+                    return false;
+                }
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e, "create new role operation failed");
+                return false;
+            }
+        }
+        public async Task<bool> DeleteRole(Guid roleId)
+        {
+            try
+            {
+                _logger.LogInformation("delete role operation start");
+                bool isSuccess = await _roleRepository.Delete(roleId);
+                if (isSuccess == true)
+                {
+                    _logger.LogInformation("delete role success");
+                    return true;
+                }
+                else
+                {
+                    _logger.LogInformation("delete role failed");
+                    return false;
+                }
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e, "delete role operation failed");
+                return false;
+            }
+        }
+        public async Task<List<DisplayRoleDTO>?> DisplayAllRole()
+        {
+            try
+            {
+                _logger.LogInformation("display all role operation start");
+                List<Role>? roles = await _roleRepository.GetAll();
+
+                if(roles == null || roles.Count == 0)
+                {
+                    _logger.LogError("roles are null");
+                    return null;
+                }
+
+                List<DisplayRoleDTO> displayRoleDTOs = roles.Select(r => new DisplayRoleDTO
+                {
+                    RoleName = r.RoleName,
+                }).ToList();
+                return displayRoleDTOs;
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e, "display role operation failed");
+                return null;
+            }
+        }
+        public async Task<bool> UpdateRole(UpdateRoleDTO updateRoleDTO)
+        {
+            try
+            {
+                _logger.LogInformation("update role operation start");
+                Role role = new Role
+                {
+                    RoleId = updateRoleDTO.RoleId,
+                    RoleName = updateRoleDTO.RoleName,
+                    UpdateAt = DateTime.UtcNow,
+                    UpdateBy = Guid.Empty
+                };
+
+                return await _roleRepository.Update(role);
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e, "update operation failed");
                 return false;
             }
         }
